feat: add user task completion command objects
Creates the API for issueing user task completion commands in the Java client.
User task completion can take a custom "action" and a Map of "variables".

Note: due to the completion REST request object expecting a proper Map for variables,
we currently only support a proper Map for the client command as well. This might extended
by follow-up issues.