feat(gateway-rest): error responses are content type problem+json
Ensure that when a REST API controller maps an Error to a Response,
the content type of the response is set to application/problem+json.
This makes it easier for REST clients to recognize when a received
response is an error. As a result, the can optimize their response
processing logic.