fix: use regular hashset to ensure it can be serialized by Kryo
Alternately we can add ImmutableCollections#Set12 and SetN to Namespaces.BASIC.
Depending on the size of the set either Set12 or SetN is used. The concrete
class also changes whether we are using Set.of() or Set.copyOf. Since there
are so many variables, it is better to use a concrete class directly to ensure
consistent serialization.