refactor: change generated client REST model classes package
Adjusts the package of REST model classes generated for the Java client from
`io.camunda.zeebe.gateway.protocol.rest` to `io.camunda.zeebe.client.protocol.rest`.

Otherwise, in integration tests, the model classes used in the gateway controllers
can be those generated for the client. This can lead to inconsistent behavior since
we use different OpenAPI generators for the components, e.g. Spring vs. HttpClient.

For example, the `UserTaskUpdateRequest` model class and its nested `Changeset` differ
between the REST gateway and the Java client. While the gateway model class is a POJO,
the client model class extends the `HashMap` class. If the latter is now used in the
controller, the deserialization of the request data puts all `Changeset` attributes
into the `HashMap` instead of mapping them to the defined properties.