feat: Create multi-instance compensation subscription on complete
Align the compensation behavior for multi-instance with other elements.

Create the compensation subscription when completing the multi-instance activity.

Don't complete or delete the multi-instance subscription on completing the multi-instance activity.

Set key of the multi-instance as the scope key of a multi-instance subprocesses.

Remove the deletion of the multi-instance subscription for now. We will implement the deleting as a follow-up for all
elements.