fix: recordstream between predicates
The between method provides a way to limit a stream to some subsequence
defined by a lower and upper bound.

It was implemented with a limit on the upper bound first, because this
returns a RecordStream on which we can use skipUntil for the lower bound
which also returns a RecordStream. This means no wrapping with the
supply method was needed.

However, if the upper bound predicate is true for any records that exist
before the lower bound, then the stream would be wrongly limited to that
first record.

I encountered this while writing some tests, where a specific event
exists on the log multiple times, once before the lower bound and once
inside the lower and upper bounds.

By using dropWhile we get a Stream, and so we need to use the supply
method to transform it back into a RecordStream before we can use the
limit method.

Note that this simplifies the implementation because when you think
about the between method, you'd expect that it first skips records until
the lower bound and only then starts checking the upper bound.