test: ensure a snapshot exist before OOD
If there is no snapshot with a compactable processed
position, we cannot compact because Replay is paused when
disk is out of space. Hence processed position would not
advance, even if exporting resumes. As a result segments
cannot be compacted. This can happen in production also,
but rare compared to going out of disk due to exporting.
This can happen if processing was explicitly paused which
resulted in accumulating many events. This is not part of
the normal functionality of zeebe. So we decided not to address
this case now.