refactor(engine): split migration tests
The MigrateProcessInstanceTest class contains 1500 lines after adding subprocess migration
tests. It is expected to grow with additions to migration logic. Because of that, we decided
to split the test class into different classes based on the component. Additionally, I added a
folder specific to migration since the number of migration test classes is also growing. That
separation is expected to add clarity for developers on migration logic capabilities.