feat: Store compensation handler instance key in subscription
Adjust the compensation behavior to store the element instance key of the compensation handler in the compensation
subscription. Use this instance key to complete the subscription that belongs to this instance of the compensation
handler.

Without the instance key of the compensation handler, we may complete a different subscription and complete the
compensation throw event of a different scope.