docs: clarify command distribution behavior
I want to refactor the CommandDistributionBehavior. So, I started to add
some additional JavaDoc to the class/methods to explain what it does and
how it works.

This pushed me to re-read our docs entry for generalized_distribution.
While reading, I saw some opportunities to improve the text:
- clarified why we need generalized_distribution
- clarified that CommandDistribution records only exist on the
  distributing partition
- added a tip to use DistributedTypedRecordProcessor

Then I came back to the JavaDoc of the CommandDistributionBehavior:
- linked to the generalized_distribution docs
- clarified that the distributeCommand method distributes to all other
  partitions.
- clarified what the distributionKey is used for as this key is very
  important to the feature. It must encode the distributing partition,
  and must be unique.