feat: accept install requests that are retried by the leader
If a leader observed a time out for an InstallRequest,
it will resend the same chunk. If the follower had
already processed the previous request successfully, it
should simply accept the request instead of rejecting
it with an out of order error. If it was actually out
of order, it will be identified during commit  as the
checksum would not match.

(cherry picked from commit be3c7d703e8ac9c84c3bf2ada4efcde1e39ba5da)