refactor: allow switching between REST and gRPC per call
To enable testing the various endpoints while we support both gRPC and
REST for the client, we introduce a toggle to specify which protocol to
use per call, as well as a setting (with property and environment
variable) to specify the default behavior on the client in general.

Additionally, general `ClientTest` tests are updated to always gRPC, as
it would be a lot of refactoring at the moment to fix that. Instead, we
added a separate `TopologyRequestRestTest` class to test the REST API.
In the future, we should "merge" these to allow for client tests to run
on both REST and gRPC where applicable.