test: stabilize flaky test by increasing timeout
This test was flaky, especially in CI, because sometimes the leader step down would be delayed by multiple seconds.
Instead of first waiting for the leader step down to occur and then checking that the request fails, I've decided
to simply increase the time we wait for the request to fail. This is because request failure can only happen on leader
step down, and waiting twice for effectively the same thing doesn't improve this test.

(cherry picked from commit d035dfedd109dfcf8608425dc1cfa556a4859352)