refactor: ensure REST API gateway property is always set
We enable/disable the REST API based on whether the embedded gateway is
enabled or not. Since the property is read from the environment and not
from the beans, the programmatic configuration via `withGatewayConfig`
is not actually used during Spring autoconfiguration. Thus we need to
set the property manually before starting the test application.