feat: allow cancellation of scheduled delayed tasks
Provides a handle to control scheduled delayed tasks and allow cancellation.
This is useful for engine checkers with more complex scheduling logic that may want to
reschedule a run, i.e. cancel a previously scheduled task and schedule a new one.
This is needed to resolve a bug in the `DueDateChecker` where we assumed that only one
execution of the task is scheduled. This probably should be a design restriction for all
engine checkers.

Rather than expose the `ScheduledTimer` from the scheduling module, we define a new interface
to not add additional dependencies from the engine to the scheduling module and to allow extension
of the `ScheduledTask` interface in the future.

Co-authored-by: Nico Korthout <nico.korthout@camunda.com>
(cherry picked from commit a5e85b0a7c1245ceffd59b6ebfe6e119f94cc28b)
(cherry picked from commit 6d6a1175264f9a48650c9c4bee01878d338c94e0)