feat: allow cancel async scheduled delayed tasks
This allows scheduled task that are supposed to run async to be cancelled too.
This is needed because the `DueDateChecker` optionally uses async scheduling and
still needs to maintain the guarantee that only one task is scheduled at a time.

Cancellation occurs after scheduling. Because of this, cancellation may have no effect
and we could actually break this guarantee still. This should be a rare case because it depends
on the delay and actor scheduling and the impact expected to be minimal and not cause further bugs.
In the case of timer expiry, breaking the guarantee may result in duplicate commands on the log
because the command cache is ineffective. But it wouldn't cause the high impact bug where
thousands of executions are scheduled at the same time.
If we see that this becomes a problem for performance still, we could consider preventing execution
after cancellation more strictly.

Co-authored-by: Nico Korthout <nico.korthout@camunda.com>
(cherry picked from commit 6792fb27d7566d178146a2bf1d17329f622a3d28)