refactor: keep track of next execution
When we schedule another run of the DueDateChecker we get returned a
ScheduledTask that can be canceled. To do so later, we need to first
store and keep track of it.

We do this by replacing the existing nextDueDate field with a new
nextExecution.

Note that on pause we continue to keep the same behavior where we only
remove our knowledge of the next execution rather than also cancel it
already. This can come in a later commit.

Co-authored-by: Ole Schönburg <ole.schoenburg@gmail.com>