refactor: remove unnecessary execute class
Execution of the due date checker is pretty simple (and has nothing to
do with TriggerEntities anymore). So, we replaced the
TriggerEntitiesTask class with a single private method execute close to
the top of the DueDateChecker class. It's close to the top, because it's
the main function of the checker.

This removed a bunch of indirection that made the class confusing to
read.

Co-authored-by: Ole Schönburg <ole.schoenburg@gmail.com>
(cherry picked from commit cc65836d9190e08ac313fb059f0b39c7e4ef1aa8)
(cherry picked from commit 3d6e6752aec0b9e6e6db15b902d33128e4a494e6)