fix: rely on cancelation instead of pause flag
The execute method previously only visited if the
shouldRescheduleChecker flag is true. However, that flag is clearly
meant to be used for rescheduling.

Note that the actor has chosen to execute this still even though the
lifecycle of the stream processor has changed. It's fine to execute one
last time, and then no longer reschedule in the schedule method. We can
allow the schedule method to make that decision.

Co-authored-by: Ole Schönburg <ole.schoenburg@gmail.com>