test: correct wrong mockito test cases
These tests depend on mockito and so they easily fail when the
implementation changes.

In this case, the behavior changed because we now require execute to be
called in other for the next call to schedule to actually schedule a new
execution.

We can also take this opportunity to assert that a previous schedule is
cancelled when we reschedule an earlier execution.

Co-authored-by: Ole Schönburg <ole.schoenburg@gmail.com>