refactor: replace initial scheduling
The initial scheduling doesn't really do anything special. We can simply
replace it with a call to schedule with -1 as dueDate.

This change will soon allow us to remove the cancelNextExecution method
and reach full threadsafety.

Note this is not a full refactor, as the initial scheduling occurs after
a 100ms delay now, while it was at ZERO duration previously. This should
not cause any strange behavior, and could actually allow the stream
processor to start up more efficiently.

Co-authored-by: Ole Schönburg <ole.schoenburg@gmail.com>