fix: avoid setting nextExecution to null
Rather we can use the due date of the nextExecution to determine whether
we need to reschedule.

This means we don't have to set the nextExecution to null in the execute
method.

As a cancel method we can simply provide a no-op for this initial
value.

Co-authored-by: Ole Schönburg <ole.schoenburg@gmail.com>
(cherry picked from commit 30c04057082b8f917180c34be34d3a720166d4f5)