fix: update should return null if unchanged
We had a bug where the AtomicUtil.update method is used, where we would
cancel the next execution if we tried to schedule a new one but didn't.

This happened because the update method always returned the previous
value. This is typical for updateAndGet like methods, but we cannot use
this constuct in our usage (specifically there is no way to check
against this case otherwise).

So, instead we decided to change the contract of this update method to
return null when the value is left unchanged. This should be easy to
reason about for users.

Co-authored-by: Ole Schönburg <ole.schoenburg@gmail.com>