fix: do not rely on clock to reschedule
Use a null value of `nextExecution` to signal that nothing is scheduled
instead of guessing if it has executed based on the current time.
Using the current time is not reliable because timers that are due now but
are slightly delayed would get cancelled unnecessarily. Additionally, this
logic would always replace the initial execution because it was scheduled
with an artificial dueDate of -1. If we scheduled something far in the
future before the initial execution occurred, we would not execute for a
long time. For example, we would not trigger timers that are already in the
state, and instead wait for the newly added timer only.

Now, when we do execute, we first set `nextExecution` to null. This allows
self-rescheduling and concurrent scheduling of new executions, e.g. when new
timers are added during processing.

While this removes us further from the goal of scheduling exactly one
execution (see comment on `execute`), it does not result in a serious issue
because it is a rare edge case that depends on a race condition between
scheduling, cancellation and execution. Every time this occurs, we may end
up with one unnecessarily scheduled execution. This resolves itself because
the next schedule will most likely not encounter the same race condition
again.

(cherry picked from commit f1960e84dbd631dacf3b9cd13a7b4253fdbcf491)