fix: avoid rescheduling for past due dates
There is a special case for due dates in the past.

Consider the following example:
- a run is scheduled for a due date 1h in the past, it is scheduled for
  100ms in the future (timerResolution)
- another run is scheduled for a due date 2h in the past, it would
  cancel the next execution (of the 1h past due date), and reschedule in
  100ms again

This is unnecessary and could lead to small issues where many
schedulings of past due dates would endlessly postpone execution.

We can avoid this by storing the time at which we schedule the next
execution and use this in the check to determine whether we need to
schedule.

We chose to inline the entire calculateDelayForNextRun method as it
this usage is slightly more comlex, and cannot easily be extracted into
a separate method in a way that reduces complexity.

Instead, we chose to clarify the names of variables (now, scheduleFor,
scheduledFor, and delay all have specific meaning).

Co-authored-by: Ole Schönburg <ole.schoenburg@gmail.com>
(cherry picked from commit dd760a3b9aba57cbdc0731623f09471b55edd4ee)