fix: schedule initial execution immediately
Instead of using the `timerResolution` as floor
and thus always scheduling for 100ms in the future, event for the initial
exection, we now special case the initial execution and schedule it
immediately.

This reverts back to the previous behaviour and unflakes the
`ProcessExecutionRandomizedPropertyTest`.
The test was flaky for timer start events because depending on actor job
scheduling order, the scheduling would sometimes happen after the clock was
moved for the test. There's currently no way to control this but we
considered adjusting the scheduling logic such that it calculates the delay
only once and the actor clock used in tests would update atomically to
mirror the normal actor clock behaviour. This seemed too much work for now
(required API changes) and this simple fix was enough to reduce the
flakiness again.

Co-authored-by: Nico Korthout <nico.korthout@camunda.com>