test: update tests to match new interceptor behavior
With the changes in interceptor behaviour, they are not invoked
if the the receiver is disconnected. To simplify the handling of
such cases, we now add a method for one way disconnect to get the
similar behaviour.

(cherry picked from commit 8165ca7b2fb1975470cd7cd6f1d71c8567bb6e54)