fix: avoid race between scheduling and execution
Because we schedule execution before updating the `nextExecution` atomic,
there was an edge case where the execution would also happen before.

This was especially likely for the initial execution which was scheduled for
`now`. In that case, we would start with `nextExecution=null`, the execution
would again set it to `null` and then the scheduling would set it to the
planned execution that already occurred. As a result, no new timers would be
scheduled.

To resolve this, we don't use null as a sentinel value and instead use fresh
instances of `None`. This allows scheduling to observe that the execution
already occurred because the CAS will see a new instance of `None` instead
of the "same" null value.

This has the drawback that there is a liveness issue when execution occurs
before scheduling is done. To reduce the likeliness, we again schedule the
initial run with a 100ms delay, making it unlikely that scheduling has to
try several times which would lead to redundant executions.