fix: ensure atomic update when not replacing
This method must ensure that the atomic reference is not updated by
another thread in between the atomic call to the replacer function and
placing it's result in the reference.

When the replacer function returns an empty optional, this method did
not verify that another thread had not replaced the reference value in
the meantime.

We're now changing that by removing the early return, and instead
compareAndSet the ref to the same value that it had at the start of the
do/while.

Note that this requires an additional check to avoid rolling back if the
replacer method returned empty but the ref was changed.