test: update clock after adjusting it
The actor clock must be updated. Typically this happens by the actor
itself in between actor jobs. When using the ControlledActorClock
directly, we must call update explicitly if there is no actor doing that
for us.

In these tests, we're using the actor clock directly and there is no
actor controlling it.