fix: jobs incidents shouldn't be resolved if retries<0
It was possible to resolve job incidents
when job retries were less or equal to 0
and to active the job again.
The check to verify that the job retries
are greater than 0 was added to fix.
We select less than 0 because of the known bug
#15437
Job incident tests were missing the update
 job retries part and this is also fixed
 now by applying job update retries command.