fix: jobs incidents should not be resolved if retries are less than 0
It was possible to resolve job incidents when job retries were less or equal to 0
and to active the job again.
The check to verify that the job retries are greater than 0 was added to fix.
We select less than 0 because of the known bug
https://github.com/camunda/zeebe/issues/15437
Job incindent tests were misssing the update job retries part and this is also
fixed now by applying job update retries command.