ci: split into separate jobs
Create new prepare / perfom job to separate of concerns.

The split allows to make more clear what each step does, and
what each parameter is used for.

For example are we able to reduce the parameters to a minimum, and
skip re-building in the perform step. This reduces the build time
from 5 minutes to 30 seconds (factor 10).

Add for interesting parts and parameters some documentation to
better understand the what and why