fix: do not short-circuit retry for force configure request
In the first attempt, both members received the force configure request,
0the current leader steps down. When the leader role is started the
comes out of force configuration and also sends the new configuration to
the follower. However, it is possible that the first request was interpreted
as failed because either the response was lost or the the pod got restarted
before it could mark it as completed. When the request is retried, the
follower that receives the request accepts its, overwrites the local
configuration with force configuration and send the request to the current
leader. The current leader sees that the membership already matches the
requested configuraiton, so it simply accepts the requests without overwriting
the local configuration. As a result, the follower remains in the force
configuration state because the leader will never re-send the normal configuration.

To fix this, we remove the optimization where the leader short-circuit the
duplicate force request. Instead, the leader always overwrites the local
configuration and go through the whole force configuration process.