fix: shutdown command might happen before soft pause command reaches all partitions.
I was not able to reproduce the issue locally, so I added debug statement
to throw the exported positions in case of error.

The potential fix added was to assert after restart that all partitions
were in fact soft paused before sending the first publish command, thus
avoiding the potential race condition.