build: symlink idea folder for zeebe
It's annoying to open the mono repo in IntelliJ when working only on
Zeebe. I prefer to work in the `zeebe` subfolder. However, that means
that all the shared idea settings are not available.

By symlinking the .idea folder users that open the zeebe subfolder in
IntelliJ will have all the same settings as the mono repo, without
having to duplicate the settings.

This change does not affect anyone that works from the mono repo in
IntelliJ.