fix: retry SetVariable on for UNAVAILABLE and RESOURCE_EXHAUSTED
SetVariable endpoint is making changes on the state of the process instance.
Therefore, it is not idempotent. If we retry setting variables in different
stages of process instance, it might provide different behaviour. Because of
that we shouldn't retry SetVariables on timeouts (DEADLINE_EXCEEDED) where
the processing of the engine can continue, so the process instance state changes.
For UNAVAILABLE and RESOURCE_EXHAUSTED, the broker will not accept further commands,
so that when we retry the SetVariables command, the process instance state will
stay the same.

(cherry picked from commit 9fd285e51575acf413d659045cc4054ded85241b)