fix: always a second entry
If there is only one entry then limit(2) will simply wait 5 seconds and
then provide a stream with just the first entry. So .getLast() then
returns that first entry, leading to the following assertion error:

java.lang.AssertionError:
Expecting actual:
2251799813685258L
to be greater than:
2251799813685258L

This error is not very understandable. We can be more explicit by
skipping the first entry and then getting the first entry. If there is
no second entry this will fail with an assertion error stating that no
entry was found.