test: ensure no leader before restarting outdated follower
The test was flaky because we were looking for a leader among m1 and m2. But
it should be m1 and m2Restarted. In the flaky case, m2Restarted was the leader.
This was not the intention of the test, because node 2 must be outdated at this
point and cannot become the leader. But in this case, the shutdowns and restarts
were so fast that leader m1 didn't get a chance to step down. So when m2 was
restarted, it immediatly got the latest log entry and was able to become the leader
immediately after the force configure is completed. To prevent this case, we now wait
until there is no leader, before restarting node 2.