test: log when we want to travel through time
My hypothesis for the flakiness of this test is that we travel through
time before the DueDateChecker's run (for the TimerDueDateChecker) is
scheduled. The scheduling happens async and uses a delay. This means
that when the scheduling happens it could occur after the clock has
already been moved. If that's the case, we do have a timer that has
exceeded it's deadline in the state, but the due date checker is not
scheduled until a later point in time that is not reached during the
test's lifetime.