refactor: allow all semantic versions
Modifies the Atomix utility `Version` to use the more complete
`SemanticVersion` utility from Zeebe. This fully complies with the
semantic version spec, and is much more lenient about adding
build-metadata and so on.

At the same time, we preserve the check on start up that the version is
parse-able at all, to avoid issues where we would run with un-parseable
versions and potentially use a snapshot which could cause
inconsistencies.