refactor: flow control keeps track of in-flight entries
Instead of passing `InFlightEntry` as `AppendListener` to the raft
layer, we keep a map of in-flight entries in `FlowControl` which is now
the `AppendListener`.

This is necessary for the next steps where we add user command flow
control, because we can't pass in `InFlightEntry` through the entire
system anymore. Instead, we'll only notify `FlowControl` about positions
that are written, committed, processed and finally exported.