refactor: rename migration preconditions
This class was named *Checker. However, checker is a core concept in the
zeebe engine implementation. Checkers periodically check the state and
append commands accordingly. This precondition checker does not do that.
We think it's better to rename this class so it doesn't contain the word
Checker to avoid confusion.

Co-authored-by: berkaycanbc <berkay.can@camunda.com>
(cherry picked from commit 48ca4e15f7e031522e6557c51307875804ba53a4)