feat: support migration of call activity
We can simply support the call activity element type as a migratable
element. There's no need to do anything to the child process instance.

We've verified that the child process instance does not have any
properties that can be updated by the migration of a call activity. Some
of the options that would change are the following, but none of these
are tracked on the log or in the state for child process instances:
- the process definition key of the parent process instance
- the bpmn process id of the parent process instance
- the version of the parent process instance
- the element id of the calling activity

Co-authored-by: berkaycanbc <berkay.can@camunda.com>
(cherry picked from commit e17a4a32d7e51e45ee26924173bf5266a413cd47)