test: allow to terminate after migrating parent
This is a reasonably interesting case to test for, as the call activity
adds a relation between two process instances that is unique to call
activities. This relation is also used on cancelation of the parent. So,
even if we currently don't need to adjust any properties of the child
process instance on migrating the parent, we should still test for this
case.