refactor: change return type
Changing "isPlaintextConnectionEnabled" return type to support null
value indicating that there is no configuration for it and avoiding
using default value when the value is actually not set

"getGrpcAddress" and "getRestAddress":
Return null if not set so default value could be set after checking new
config. This done because we check the old config first and only then -
the new one.