fix: warn on ignored scheduled task
If the scheduling service is not yet up-and-running and it is used to
schedule a task, the task is simply ignored and nothing is scheduled.

While that could make sense in some cases, it could also hide a bug. It
therefore makes sense to log such cases at warn level.