feat: schedule timer at timestamp
This adds a new feature to the ActorControl: scheduling a timer task at
a specific timestamp.

Note that due to the way the ActorTimerQueue is implemented
(DeadlineTimerWheel), there is no guarantee that the task is run AT the
timestamp.

(cherry picked from commit 05a3b3fc763947b0dcadc7dae74aa2de40d64ae3)