fix: schedule due date checker at timestamp
This is to unflake tests that expect to trigger a timer through the due
date checker after traveling through time.

This was unstable because the due date checker was scheduled with a
delay, where this delay is used to determine an actual deadline at a
later point in time (during scheduling of the TimerSubscription by the
actor). The unstability arose because the task can be scheduled async
which means that the deadline was calculated async as well, so it could
happen that the deadline was determined after the test had already
travelled through time.

This problem is resolved by using the new runAt method, which allows
scheduling a task at a specific timestamp. When the scheduling is
postponed until after the time traveling, the task will trigger
immediately if the timestamp is now in the past.