test: add tests for migrating a child process instance
Three different test cases are added. First one is testing if the
migration of a child process instance can be successfully completed.
Second, the test verifies if the flow can continue as expected in
both parent and child process instances. Third, we add test if
cancellation of a parent process instance still works as expected.
It is basically the same test cases as we have for call activity
migration but with migrating child process instance this time.