test: add tests for migrating a child process instance
Three different test cases are added. First one is testing if the
migration of a child process instance can be successfully completed.
Second, the test verifies if the flow can continue as expected in
both parent and child process instances. Third, we add test if
cancellation of a parent process instance still works as expected.
It is basically the same test cases as we have for call activity
migration but with migrating child process instance this time.

(cherry picked from commit ac9c59550d9231f78db35a24c28ca146a8970036)