refactor: support migrating child process instances
The precondition to prevent migrating child process instances is removed.
There's no need to do anything to the child process instance specifically.

We've verified that the call activity that is referencing the migrated child
process instance does not have any properties that needs to be updated on migration.
Following fields can be changed during the migration of child process instance:
- the process definition key of the child process instance
- the bpmn process id of the child process instance
- the version of the child process instance
- the element id of the active element inside the child process instance

Checked following classes to see if anything needs to be updated for the calling
call activity:
`ProcessInstanceRecord` -> no data kept for child process instance
`ElementInstance` -> only called child instance's key is kept but it will not change
during the migration
`DBElementInstanceState` -> parent and child instances are connected through instance
keys that will not change during the migration of child process instance

Therefore, there is no need to update call activity neither in log stream nor in the
state.